fogen: protect init from self overriden functions
if LIBFO_INIT_FILE is set, user overrides fgets() and installs
failing point in fgets(), once we initialize failing point for
fgets() our call to fgets() might fail by triggering failing point.

This is because when fgets() is overriden, when we call fgets()
we go via our overriden routing and not original one. This applies
to all functions used in fo_init().

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>